auth-backend/requirements.txt
# fastapi & dependencis
fastapi
uvicorn
starlette
pydantic
pydantic[email] # for email validation
# sqlalchemy db orm
sqlalchemy
# `.env`ファイルから環境変数を読み込むためのライブラリ
python-dotenv
# poregresql library
psycopg2-binary
# formatter
autoflake
isort
black
# テスト用
pytest
pytest-timeout